+++ xlb/__init__.py
@@ -2,7 +2,7 @@
 from xlb.compute_backend import ComputeBackend as ComputeBackend
 from xlb.precision_policy import PrecisionPolicy as PrecisionPolicy, Precision as Precision
 from xlb.physics_type import PhysicsType as PhysicsType
-from xlb.mres_perf_ptimization_type import MresPerfOptimizationType as MresPerfOptimizationType
+from xlb.mres_perf_optimization_type import MresPerfOptimizationType as MresPerfOptimizationType
 
 # Config
 from .default_config import init as init, DefaultConfig as DefaultConfig
