+++ xlb/helper/simulation_manager.py
@@ -2,7 +2,7 @@ import neon
 import warp as wp
 from xlb.operator.stepper import MultiresIncompressibleNavierStokesStepper
 from xlb.operator.macroscopic import MultiresMacroscopic
-from xlb.mres_perf_ptimization_type import MresPerfOptimizationType
+from xlb.mres_perf_optimization_type import MresPerfOptimizationType
 
 
 class MultiresSimulationManager(MultiresIncompressibleNavierStokesStepper):
@@ -106,7 +106,7 @@ class MultiresSimulationManager(MultiresIncompressibleNavierStokesStepper):
                 timestep=0,
             )
 
-        def recursion_fused_finest(level, app, is_self_f1_the_explosion_src_field, is_self_f1_the_coalescence_dst_field):
+        def recursion_fused_finest(level, app):
             if level < 0:
                 return
 
@@ -123,8 +123,7 @@ class MultiresSimulationManager(MultiresIncompressibleNavierStokesStepper):
                     missing_mask=self.missing_mask,
                     omega=self.omega,
                     timestep=0,
-                    is_f1_the_explosion_src_field=is_self_f1_the_explosion_src_field,
-                    is_f1_the_coalescence_dst_field=is_self_f1_the_coalescence_dst_field,
+                    is_f1_the_explosion_src_field=True,
                 )
                 self.add_to_app(
                     app=app,
@@ -136,8 +135,7 @@ class MultiresSimulationManager(MultiresIncompressibleNavierStokesStepper):
                     missing_mask=self.missing_mask,
                     omega=self.omega,
                     timestep=0,
-                    is_f1_the_explosion_src_field=not is_self_f1_the_explosion_src_field,
-                    is_f1_the_coalescence_dst_field=not is_self_f1_the_coalescence_dst_field,
+                    is_f1_the_explosion_src_field=False,
                 )
                 return
 
@@ -161,10 +159,10 @@ class MultiresSimulationManager(MultiresIncompressibleNavierStokesStepper):
             # so is_self_f1_the_explosion_src_field is True
 
             if level - 1 == 0:
-                recursion_fused_finest(level - 1, app, is_self_f1_the_explosion_src_field=True, is_self_f1_the_coalescence_dst_field=True)
+                recursion_fused_finest(level - 1, app)
             else:
-                recursion_fused_finest(level - 1, app, is_self_f1_the_explosion_src_field=None, is_self_f1_the_coalescence_dst_field=None)
-                recursion_fused_finest(level - 1, app, is_self_f1_the_explosion_src_field=None, is_self_f1_the_coalescence_dst_field=None)
+                recursion_fused_finest(level - 1, app)
+                recursion_fused_finest(level - 1, app)
             # Important: swapping of f_0 and f_1 is done here
             print(f"RECURSION Level {level}, stream_coarse_step_ABC")
             self.add_to_app(
@@ -182,9 +180,7 @@ class MultiresSimulationManager(MultiresIncompressibleNavierStokesStepper):
         if self.mres_perf_opt == MresPerfOptimizationType.NAIVE_COLLIDE_STREAM:
             recursion_reference(self.count_levels - 1, app=self.app)
         elif self.mres_perf_opt == MresPerfOptimizationType.FUSION_AT_FINEST:
-            recursion_fused_finest(
-                self.count_levels - 1, app=self.app, is_self_f1_the_coalescence_dst_field=None, is_self_f1_the_explosion_src_field=None
-            )
+            recursion_fused_finest(self.count_levels - 1, app=self.app)
         else:
             raise ValueError(f"Unknown optimization level: {self.opt_level}")
